\name{warp.time}
\alias{warp.time}
\title{Calculate warped time points}
\description{Given a vector of time points and a warping function,
  calculate the new time points.}
\usage{
warp.time(tp, coef)
}
\arguments{
  \item{tp}{Vector of time points}
  \item{coef}{Warping coefficients}
}
\value{A vector of warped time points}
\author{Ron Wehrens}
\examples{
}
\keyword{manip}
