\name{gaschrom}
\docType{data}
\alias{gaschrom}
\title{16 calibration GC traces}
\description{
  This dataset contains 16 calibration GC traces, measured at 5,000 time
  points.
}
\usage{
data(gaschrom)
}
\source{Claire Boucon, Unilever}
\references{
  Eilers, P.H.C. Parametric Time Warping. Anal. Chem., 2004, 76, 404 - 411
}
\keyword{datasets}
