\name{lcms}
\docType{data}
\alias{lcms}
\alias{mz}
\alias{time}
\title{Parts of 3 proteomic LC-MS samples}
\description{
  The data set consists of a 100 x 2000 x 3 array \code{lcms}, a vector \code{time} of length 2000 and a vector \code{mz} of length 100. The LC-MS data in the array are a subset of a larger set measured on a tryptic digest of E. coli proteins.
}
\usage{
data(lcms)
}
\source{Nijmegen Proteomics Facility, Department of Laboratory Medicine, Radboud University Nijmegen Medical Centre
}
\references{
  Bloemberg, T.G., et al. (2010) "Improved parametric time warping for Proteomics", Chemometrics and Intelligent Laboratory Systems, \bold{104} (1), 65 -- 74.
}
\keyword{datasets}
