\name{ptw-package}
\alias{ptw-package}
\docType{package}
\title{
\packageTitle{ptw}
}
\description{
\packageDescription{ptw}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{ptw}
}
\author{
\packageAuthor{ptw}

Maintainer: \packageMaintainer{ptw}
}
\references{
@Article{Bloemberg2010,
  title = {Improved Parametric Time Warping for Proteomics},
  author = {Tom G. Bloemberg and Jan Gerretzen and Hans J. P. Wouters and Jolein Gloerich and Maurice {van Dael} and Hans J. C. T. Wessels and Lambert P. {van den Heuvel} and Paul H. C. Eilers and Lutgarde M. C. Buydens and Ron Wehrens},
  year = {2010},
  journal = {Chemometrics and Intelligent Laboratory Systems},
  volume = {104},
  number = {1},
  pages = {65-74},
}

Article{Wehrens2015,
  title = {Fast parametric warping of peak lists},
  author = {Ron Wehrens and Tom G. Bloemberg and Paul H. C. Eilers},
  year = {2015},
  volume = {31},
  pages = {3063--3065},
  journal = {Bioinformatics},
  doi = {10.1093/bioinformatics/btv299},
}
}
\keyword{ package }
